+++ src/idlib/hashing/CRC32.h
@@ -2,6 +2,8 @@
 #ifndef __CRC32_H__
 #define __CRC32_H__
 
+#include <stdint.h>
+
 /*
 ===============================================================================
 
@@ -11,9 +13,9 @@
 ===============================================================================
 */
 
-void CRC32_InitChecksum( unsigned long &crcvalue );
-void CRC32_UpdateChecksum( unsigned long &crcvalue, const void *data, int length );
-void CRC32_FinishChecksum( unsigned long &crcvalue );
-unsigned long CRC32_BlockChecksum( const void *data, int length );
+void CRC32_InitChecksum( uint32_t &crcvalue );
+void CRC32_UpdateChecksum( uint32_t &crcvalue, const void *data, int length );
+void CRC32_FinishChecksum( uint32_t &crcvalue );
+uint32_t CRC32_BlockChecksum( const void *data, int length );
 
 #endif /* !__CRC32_H__ */
