+++ src/idlib/hashing/MD5.cpp
@@ -257,7 +257,7 @@ void MD5_Final( MD5_CTX *ctx, unsigned char digest[16] ) {
 MD5_BlockChecksum
 ===============
 */
-unsigned long MD5_BlockChecksum( const void *data, int length ) {
+uint32_t MD5_BlockChecksum( const void *data, int length ) {
 	static_assert( sizeof( unsigned int ) == 4, "MD5 requires 32-bit digest words" );
 	unsigned int	digest[4];
 	unsigned int	val;
