+++ src/idlib/hashing/MD5.h
@@ -2,6 +2,8 @@
 #ifndef __MD5_H__
 #define __MD5_H__
 
+#include <stdint.h>
+
 /*
 ===============================================================================
 
@@ -11,6 +13,6 @@
 ===============================================================================
 */
 
-unsigned long MD5_BlockChecksum( const void *data, int length );
+uint32_t MD5_BlockChecksum( const void *data, int length );
 
 #endif /* !__MD5_H__ */
